- 인덱스 설계: (parent_folder_id, ID, file_id) 복합 인덱스로 완료   - 옵티마이저 이슈: ORDER BY 비용 오판을 USE INDEX 힌트로 해결   - 페이지네이션: all_file_list_paged로 Keyset 방식 구현 완료, 기존 all_file_list 호출부는 영향 없음

+++ UI/overcloud/overcloud/Views/AccountDetailView.xaml.cs
@@ -93,8 +93,8 @@ namespace overcloud.Views
             // “All” 필터일 때 전체 합계 행 추가
             if (filter == "All")
             {
-                ulong totalSum = (ulong)grouped.Sum(g => g.Total);
-                ulong usedSum = (ulong)grouped.Sum(g => g.Used);
+                double totalSum = grouped.Sum(g => g.Total);
+                double usedSum = grouped.Sum(g => g.Used);
                 items.Insert(0, new UsageItemViewModel
                 {
                     DriveName = "Total",
